Move ToastUtils to androidshared
