+++ collect_app/src/main/java/org/odk/collect/android/activities/CaptureSelfieActivity.java
@@ -22,7 +22,7 @@ import android.view.WindowManager;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.fragments.Camera2Fragment;
-import org.odk.collect.android.utilities.ToastUtils;
+import org.odk.collect.androidshared.utils.ToastUtils;
 
 public class CaptureSelfieActivity extends CollectAbstractActivity {
