+++ collect_app/src/main/java/org/odk/collect/android/configure/qr/QRCodeMenuDelegate.java
@@ -13,7 +13,7 @@ import org.odk.collect.android.preferences.source.SettingsProvider;
 import org.odk.collect.android.utilities.ActivityAvailability;
 import org.odk.collect.android.utilities.FileProvider;
 import org.odk.collect.android.utilities.MenuDelegate;
-import org.odk.collect.android.utilities.ToastUtils;
+import org.odk.collect.androidshared.utils.ToastUtils;
 import org.odk.collect.async.Scheduler;
 
 import timber.log.Timber;
