+++ collect_app/src/main/java/org/odk/collect/android/fragments/BarCodeScannerFragment.java
@@ -37,7 +37,7 @@ import org.odk.collect.android.R;
 import org.odk.collect.android.utilities.Appearances;
 import org.odk.collect.android.utilities.CameraUtils;
 import org.odk.collect.android.utilities.CodeCaptureManagerFactory;
-import org.odk.collect.android.utilities.ToastUtils;
+import org.odk.collect.androidshared.utils.ToastUtils;
 import org.odk.collect.android.views.BarcodeViewDecoder;
 
 import java.io.IOException;
