+++ collect_app/src/main/java/org/odk/collect/android/geo/GoogleMapConfigurator.java
@@ -16,7 +16,7 @@ import org.odk.collect.android.R;
 import org.odk.collect.android.geo.MbtilesFile.LayerType;
 import org.odk.collect.android.preferences.PrefUtils;
 import org.odk.collect.android.utilities.PlayServicesChecker;
-import org.odk.collect.android.utilities.ToastUtils;
+import org.odk.collect.androidshared.utils.ToastUtils;
 import org.odk.collect.shared.Settings;
 
 import java.io.File;
