+++ collect_app/src/main/java/org/odk/collect/android/projects/QrCodeProjectCreatorDialog.kt
@@ -30,9 +30,9 @@ import org.odk.collect.android.utilities.ActivityAvailability
 import org.odk.collect.android.utilities.CodeCaptureManagerFactory
 import org.odk.collect.android.utilities.CompressionUtils
 import org.odk.collect.android.utilities.DialogUtils
-import org.odk.collect.android.utilities.ToastUtils
-import org.odk.collect.android.utilities.ToastUtils.showShortToast
 import org.odk.collect.android.views.BarcodeViewDecoder
+import org.odk.collect.androidshared.utils.ToastUtils
+import org.odk.collect.androidshared.utils.ToastUtils.showShortToast
 import org.odk.collect.material.MaterialFullScreenDialogFragment
 import org.odk.collect.projects.ProjectsRepository
 import timber.log.Timber
