+++ collect_app/src/main/java/org/odk/collect/android/utilities/MediaUtils.java
@@ -23,6 +23,7 @@ import android.net.Uri;
 
 import org.odk.collect.android.BuildConfig;
 import org.odk.collect.android.R;
+import org.odk.collect.androidshared.utils.ToastUtils;
 
 import java.io.File;
